Refactoring universe model to support warm up
